+++ package.json
@@ -42,6 +42,5 @@
     "prettier": "^3.4.2",
     "vite": "^6.0.3",
     "vitest": "^2.1.8"
-  },
-  "packageManager": "pnpm@9.15.1+sha512.1acb565e6193efbebda772702950469150cf12bcc764262e7587e71d19dc98a423dff9536e57ea44c49bdf790ff694e83c27be5faa23d67e0c033b583be4bfcf"
+  }
 }
