+++ src/utils/Islogin.js
@@ -1,0 +1,6 @@
+const isLoggedIn = () => {
+  const user = localStorage.getItem("user");
+  return user ? true : false;
+};
+
+export default isLoggedIn;
