+++ src/utils/router.js
@@ -1,0 +1,27 @@
+import MainPage from "../pages/MainPage.jsx";
+import ProfilePage from "../pages/ProfilePage.jsx";
+import LoginPage from "../pages/LoginPage.jsx";
+import ErrorPage from "../pages/ErrorPage.jsx";
+
+const routes = {
+  "/": MainPage,
+  "/profile": ProfilePage,
+  "/login": LoginPage,
+  "*": ErrorPage,
+};
+
+export const Router = {
+  //   RouterType: "basic",
+  render: () => {
+    const root = document.getElementById("root");
+    let path;
+    // if (Router.RouterType === "basic") {
+    path = window.location.pathname;
+    // }
+    console.log("path", path);
+
+    const Component = routes[path] || routes["*"];
+    root.innerHTML = Component();
+    console.log("root.inner", root.innerHTML);
+  },
+};
